requirements.txt
annotated-types             >=0.7.0
anyio                       >=4.9.0
asyncpg                     >=0.30.0
click                       >=8.1.8
colorama                    >=0.4.6
databases                   >=0.9.0
fastapi                     >=0.115.12
greenlet                    >=3.2.1
h11                         >=0.16.0
idna                        >=3.10
pip                         >=25.1.1
pydantic                    >=2.11.4
pydantic_core               >=2.33.2
pydantic-settings           >=2.0
sniffio                     >=1.3.1
SQLAlchemy                  >=2.0.40
starlette                   >=0.46.2
typing_extensions           >=4.13.2
typing-inspection           >=0.4.0
uvicorn                     >=0.34.2
python-multipart            >=0.0.5
python-dateutil             >=2.8.2
passlib[bcrypt]             >=1.7.4
python-jose[cryptography]   >=3.3.0
cryptography                >=45.0.2